src/requirements.txt
discord.py==1.5.1
webdriver-manager==3.2.2
tda-api==0.7.1
selenium==3.141.0
requests==2.25.1
xmltodict==0.12.0